Renamed resource files and folders to lens
