+++ library/lens/mod.ts
@@ -1,0 +1,1 @@
+export { createResource, Resource } from "./lens.ts";
