--- library/resource/mod.ts
@@ -1,1 +1,0 @@
-export { createResource, Resource } from "./resource.ts";
