+++ mod.ts
@@ -4,6 +4,6 @@ export { setDefaultContext, setDefaultEngine } from "./library/global.ts";
 
 export { type SchemaInterface } from "./library/schema/mod.ts";
 
-export { createResource, type Resource } from "./library/resource/mod.ts";
+export { createResource, type Resource } from "./library/lens/mod.ts";
 
 export { createNamespace } from "./library/namespaces/namespace.ts";
